refactor(tools): 内联 ToolsContent 并清理频道折叠箭头和依赖
